Common: rewrote stream classes hierarchy
IStream base interface now declares only two generic methods for reading and writing: read/write a byte and read/write a buffer.
CAGSStream is renamed to CDataStream and is considered a base stream for working with binary data; it provides a bunch of helper functions for easier reading and writing basic types, generic implementations for reading and writing null-terminated string, and support for automatic endianess conversion.
CFileStream is a descendant of CDataStream.
CAlignedStream remains a descendant of IStream, and wraps CDataStream. It gained ownership over underlying stream, but also provides a method to release from control without deletion.