Common, Engine: AGS uses stream classes
One important thing left is to develop a wrapper over allegro's PACKFILE.

A care must be taken when replacing old fgetc and getc with new stream methods. Most *getc calls were automatically replaced with ReadInt8().
While usually it does not make much difference if ReadByte() or ReadInt8() is used, there are few cases where AGS code expects exactly integer as a return type to be able to detect EOF (-1), in which case ReadByte() should be called. I found two such cases and manually fixed them.
Other potentially dangerous case is when an unsigned byte value that may be >127 is supposed to be stored in signed integer (for any random reason). I was not able to detect those yet, the future tests will show if such cases exist.