Script: reimplemented runtime code fixups
CodeHelper concept seems a bit excessive for AGS script interpreter. It saves some memory (array of helpers is about 1.5-2.5 times less than code array, in average) but adds a risk of slowing interpreter down when running script with extensive script function calls. Not sure how much that could be a problem, but it's not like the extra array of chars is the problem on its own.
Besides, it could be a good idea to move runtime fixups to ccScript class and create them from "normal" fixups when script is loaded in game, then share same fixups array between all instances, similarly as we share the code array now. This will save memory even more, especially if we delete original fixups arrays after runtime fixups are created.