Initialize local variable in LZW algorythm
There are cases when the 'c' local variable of insert() function is used in condition while being uninitialized.
This is also reproducible in original CJ's SVN source.
In older Editor this was ignored, but now it causes runtime check failure; maybe related to changes in project settings, or building process, or stack contents (could be related to how MSVS debug system makes this check).

In any case, this deserves some fix.