Common, Engine: Graphics class
Graphics class is a step towards being more conformant with new Allegro 5 concept, according to which all the drawing is made to the preset target, as opposed to Allegro 4, where the target is defined as an argument for every drawing operation.
As of this commit, there are many cases of excessive graphics object manipulation, which reveal lack of optimization in original code. Those should be refactored in the following updates.