Common: fixed loading of compressed sprite file by SpriteCache
This bug seem to have existed for quite some time already. The only reason it was not noticed is that when game is compiled, a sprite index is included as well, which provides a table of offsets; therefore the sprite file was never fully read. The problem will arise only if the sprite index table is missing, or ignored for some reason.