Fixed issue raised by DKH: "BUG REPORT: Importing a sprite, creating a room object, clicking the "..." in the image property, picking the imported sprite doesn't actually change the image property. It leaves it at the default value (0) and hides the object properties in a rather weird fashion."
This is because when choosing a sprite, when closing the sprite selector we
added a command to refresh the sprite manager (another bug fix),
which caused the sprite manager to be updated with an empty selection
which did a reset on the property grid, and that in turn cancelled the
set of the sprite.
The fix is to only reset the property grid if what selected already is a sprite.