Script: revert StaticArrays to export direct pointer to C-array
There are cases when the script interpreter must compare pointer to object and pointer to array element, which makes it unnecessarily complicated to have an Array<T> wrapper address registered for array script symbol.
Besides, obviously, current code will still work only as far as array is linear (because of pointer math when accessing array elements in script) and never reallocated (because explicitly registered symbols will end pointing to wrong memory addresses).

What could be a possible solution to this: change the meaning of pointer (by which the item is registered) from memory address to conventional uid, which value = address of array + linear offset to element, which it would have if the array was linear. In this case we could remake linear object arrays into any other sort, like array of pointers, binary tree or hash map, while keeping compatibility with old scripts.