Android: Allow alternate default save folder.
If the directory containing the game file is non-writable the engine
will attempt to use the directory given as the base directory instead.
This is the "directory" parameter used when starting the AgsEngine
activity, and is only utilized if the default location is inaccessible.
This allows reading the game files from a read-only location while
saving to external storage.