Engine: fixed savedgame restoration with missing audio resource
Number of deprecated script functions rely on 'current_music_type' parameter. If the game was saved with music playing on one of the channels, and the same game was later restored with missing audio resource (e.g. no VOX pack), then the corresponding music will not be loaded and no channel created, but the 'current_music_type' parameter was kept unmodified, causing inconsistence, and - in particular cases - crashing the game.
This commit resolves the issue.