Android: Java library improvements, add user API
Added IAgsApp interface for user apps to override in-game events.
Currently allows in-game menu to be specified, and the following events
to be overridden: onMenuKeyPressed, onBackKeyPressed, onKeyboardEvent,
onInGameMenuItemSelected. Keyboard events are passed using the KeyCode
enum which joins the Android key code ID with the Unicode character.
Built-in menu renamed to "default_ingame" to allow apps to use the name
"ingame".