Editor: Changed Scripts class to proper collection of ScriptAndHeader objects.
The previous Scripts class was a strange amalgamation of raw Script
objects and ScriptAndHeader objects. This created an inconsistent
interface when dealing with the collection. This simply changes the
collection so that it only deals with ScriptAndHeader objects. Note that
this may break existing editor plugins. If any issues are reported
additional action may be required to approximate the removed/replaced
functions.