Global speech animation delay support
Adds Speech.UseGlobalSpeechAnimationDelay and
Speech.GlobalSpeechAnimationDelay to override
Character.SpeechAnimationDelay. The option in the General Settings pane
of the editor has been removed as this is no longer a legacy feature.
The default setting remains to use per-character delays.

(NOTE: The previous legacy implementation was never applied to
Sierra-style speech, while the current implementation does and uses the
same internal variables. This should not be considered to break
backwards compatibility as games using Sierra-style speech had no valid
reason to make use of the legacy setting. Sierra-style speech supports
changing this setting as evidenced by the per-character setting.)