Engine: fixed MYMOD::get_pos_ms()
MYMOD::get_pos_ms used irrelevant data to calculate current clip's playing time.

The problem is related to MYMOD::get_pos returning not the actual position, but the MOD/XM pattern "order" index, which, in turn, was a way to support backward-compatible behavior of the AudioChannel.Position property.

This might need a redesign in the future.