Engine: replaced RawFileOutputTarget with LogFile class
When created LogFile may open file right away or delay doing this. In the latter case it will buffer output up to certain size limit. When told to open the file, it will first flush its buffer. This allows to log events even before the log path is decided (for example, before or during reading configuration and/or parsing command line).