Fix for pane doesn't open after closing it by clicking the 'x' on top right.
The difference between clicking the 'x' and by clicking close in the context menu, is that for some reason clicking the 'x' fires the dock state changed event once, while closing via the context menu fires the event twice.
The code unknowingly relied on the event being fired twice since the IsShowing property of the docking container was not marked as false after showing the container was completed.
This was fixed.