Vsync for Direct3D
Direct3D requires that vsync be initialised when the game is started.
I've added a configuration option to winsetup.exe that determines
whether vsync is active when the game starts. In DirectX 5 mode, this
can be toggled using the System.VSync property. In Direct3D, the
property is now read-only.