Hack for navigating the document tabs with Ctrl+Tab
This branch fixes this bug here:
http://www.adventuregamestudio.co.uk/forums/index.php?issue=475.0 -- but
it does so in a kind of hacky way. When keyboard commands are issued,
_dockPanel.ActiveContent, _dockPanel.ActiveDocument, etc. are *always*
set to the old active tab's value the first time Ctrl+Tab is pressed.
This code catches that case and finds the *real* active tab by querying
the panes to find the one with the focus.