Fixed undo history deleted when changing dock state.
The problem was that the scintilla editor is disposed by the docking suite whenever there's a dock change to the document. Until now I solved it by reinitializing scintilla when that happens, but indeed the undo history is deleted...
I searched to see if there's another project that uses scintilla with the docking suite together, and found QuickSharp. And they solved it by doing a win api call on dispose that actually cancels the dispose.
I did the same and removed the reinit code (and related hacks), and it seems to work nicely. Thank you QuickSharp!