AGS.Native: set loaded game version on 2.72 project import
This lets some data reading routines to know actual format version they are loading. We switch back after game is imported to let drawing routines "see" latest game version.
In general this is indeed a hack, but the AGS.Native code is still such a mess, the proper solution would require to rewrite much of its contents, which is beyond scope of this fix.