Common: extended WFN Font support
This implements WFNFont class which is used to store font data in a structured way other than in previously used buffer of chars. This not only improves code readability and reusability, but also makes it possible to access character data directly without using lookup table and endianess corrections.

Additionally, this adds support for extended characters (those with codes 128 - 255).