Editor: Rebuild all when the debug state changes
This forces recompilation of all files if the debug state differs from
the one previously used to compile the project. Old projects will be
fully rebuilt the first time they're saved+compiled with this new
setting. This ensures that the first rebuild after opening an old
project will always be correct. A side effect of this commit is that the
/compile command line switch will now save the project if it is fully
rebuilt. I don't see this as particularly harmful. Every other way to
compile the project was saving it, so if anything this homogenises
things a bit.