Reimplemented letterboxed-by-design game mode.
Previously I misunderstood the meaning of the letterbox option in the game properties. It appears it was used to create letterbox mode as a part of game design.
This letterbox-by-design is now applied simply by making AGS pass higher resolution to the graphics mode initialization procedure. This works only for 320x200 and 640x400 games (they appear as 320x240 and 640x480 respectedly, with black horizontal borders).

This feature itself was designed in a rather awkward manner, and its future is unclear, but we leave it be for compatibility with older game projects.

This technically reverts commits:
91513cd9bb0b831d865f5b1553cacd35a86fc759
078ec5e6e3030c489c96a5b05d3a9529b97c36e9