Script: SCMD_CALLOBJ permits arguments of 'plugin argument' type
This fix adds missing permission for runtime values of "plugin argument" type to be passed to the function, expecting object pointer. "Plugin argument" is technically a 32-bit value received from plugin, that may be cast to a pointer (e.g. C-string). This is not safe on 64-bit platforms, but there's nothing we can do at the moment (plugin interface must be completely rewritten to achieve compatibility).