Script API: Implemented Mouse.GetModeView method
- Added GetModeView, similar to GetModeGraphic. GetModeView returns the
view assigned to the cursor or 0 for no view
- The view of a cursor was previously stored as -1 if the cursor wasn't
animated. Broke out the "-1 check" for view animation into a separate
flag (MCF_ANIMATE)
- For old game files, the behaviour of ChangeModeView should be the
same. It was possible to toggle whether a cursor was animated by setting
its view to 0 or a non-0 value. For new projects, this will only be true
if the cursor was set to Animated in the editor