Common: improved the GUI visibility mode representation
The GUIMain::on variable was used to define both GUI visibility and allowance to be shown.
This commit removes GUIMain::on variable and replaces it with private _visibility member, one setter and three getters which detect particular visibility modes.

GUIMain::on could have following values:
* -1 == hidden by command;
*  0 == completely disabled;
*  1 == shown by command

Now replaced with respected getters:
* IsConcealed();
* IsOff();
* IsVisible();