Engine: do not auto Unlock mutex on destruction
Mutex wrapper class should never automatically lock or unlock itself; it is not its task to control its own state like this.
There's not much sense in calling Unlock() in Mutex's wrapper's destructor, because if it was not unlocked in time, program will probably get in a deadlock anyway. On other hand, trying to unlock non-locked mutex may cause errors on some systems.