Scripting: Fix segfault when compiling functions returning String[]
The variables inside a dynamic string array have a size of 0, causing
another lookup to take place on the return type. The compiler was
treating them as though they were references to forward-declared structs
types with a previously unknown size. This would be fine under normal
circumstances, but the STYPE_POINTER and STYPE_DYNARRAY flags were
causing the lookup to occur well outside the boundaries of the symbol
size array. Stripping these flags on the lookup allows this lookup
(albeit somewhat unnecessary) to find the correct size of the variables
in a string array.