Scripting: Implemented System.RuntimeInfo
Implemented a read-only `System.RuntimeInfo` String property that
returns the information displayed when pressing Ctrl+V in a game. The
debug implementation of `debug(1, 0)` still works in debug mode.
Internally, it just calls `System_GetRuntimeInfo` and displays the info.
The `System.RuntimeInfo` property exists in debug or release mode. The
user may implement the display of this info however s/he likes.