Engine: temp fix for OGG & MP3 voice buffer underrun
This is a hotfix for voice clip stopping playing on rare occasion, when the OGG or MP3 decoder encounters buffer underrun.

AGS uses ALOGG and ALMP3 libs as intermediate between itself and OGG/MP3 decoders. Unfortunately, it seems that these libraries do not provide means to deal with situation like that properly. When streaming, instead of pulling the data from source, the simply wait when the user would push some data into their buffers and polls, in contrary to how underlying decoders work, making it impossible to interfere in time if the buffer underrun was detected by the decoder.

This fix only forces the playback to continue. A drawback is that a tiny chunk of audio data is dropped, making very short, but noticeable pause in speech.

The proper solution would be to upgrade ALOGG and ALMP3 libraries to the latest version with more clear streaming API. This would require to change how AGS works with them, so I am leaving this for the future major version (3.4.0 or higher).