Scripting: Only look up final struct sizes for functions
This was supposed to be a check for functions with prototypes inside a
managed struct that returned a pointer to that struct, e.g.
```
managed struct myStruct
{
import myStruct *getOne();
int x;
};
```
The struct size is unknown at the time of the `getOne()` declaration. At
the time of parsing the function body, the check was supposed to grab
the real struct size from the return type of the function.
Unfortunately, this was occurring with lots of symbols, not just
function bodies. Those other symbols don't really have return types (or
even an initialised funcparamtypes array for that matter).