Scripting: Write correct loop increment info
Some of the info written when inserting the loop increment chunk
incorrectly referred to nested_level, which may or may not have been
initialised. The correct location is loop_level (as can be seen in the
rest of the function call. This was a typo and/or bad copy+paste on my
part. It's possible I originally had nested_level and changed some
variables but not others on that line.