Engine: fixed Hq2x and 3x algorithms for smaller input bitmap sizes
This makes Hq2x and 3x algorithm work properly with input bitmap sizes smaller than half and one third size of the output surface, respectively.
The problem was that it did not expect the output scanline be longer than required to get all the image, and therefore did not skip bytes remaining till the next scanline, causing parts of image to appear "misplaced" on screen.