Scripting: Permit static method calls as array indices
The previous validation for expressions within an array was a little too
strict. This adds another exception (in addition to commas) for struct
types. Also added a break if the struct isn't followed by a dot. As type
conversion won't throw an error, this ensures we don't get a static
non-property access error.