Absolute Paths for Resources
Importing sprites, audio and fonts would always store absolute file path in the property "SourceFileName". It now stores relative path if the file is found inside the Game Project, or absolute path if it is fount outside of the game project. This is useful if the the game project is moved to another place on a local computer, or if the game project is shared between multiple collaberators on each their private computer.

Examples:
  Relative Path: Resources/Sprites/MySprite.bmp
  Absloute Path: D:/MyGameProject/Resources/Sprites/MySprite.bmp