Engine: correct check for disabled sound in legacy sound & music API
This correctly checks for audio clip type before starting playback to know if related sound driver is currently set up.
Previously all sounds are assumed to be of digital types and all music - MIDI type.

Note, this commit tries to mimic previously existing behavior that stopped currently playing clip only if the new sound fails to load, but not if the sound type is disabled.