Scripting: Improve switch() statement
This is an improved implementation of switch(). The previous
implementation would rewrite the chunk of code responsible for
calculating the switch variable for each case. Not only was this brittle
and buggy, it was also incorrect behaviour. If switching on a function
call, it would call the function for each case statement. This
implementation does the switch() part once and writes/reads the result
from the stack.