Editor: Mouse flickering in the code editor
Certain meesages of the type WM_SETCURSOR in Scintilla code editor was not forwarded to the underlying logic located in ScintillaWin.cxx, resulting in a rapid mouse flickering when moving the mouse over the code editor.
https://github.com/adventuregamestudio/ags/blob/master/Editor/scintilla/src/ScintillaWin.cxx

I've been unable to pin point exactly why some messages are not forwarded to ScintillaWin, however the documentation for DefWindowProc states that "This function ensures that every message is processed.", and it appears to fix the issue.
https://msdn.microsoft.com/en-us/library/windows/desktop/ms633572(v=vs.85).aspx