Engine: simplified virtual AGSPlatformDriver::WriteStdOut()
This makes virtual (and platform-specific) WriteStdOut() method to have only one string parameter which is supposed to be directly written to the std output as-is, without formatting it further.
String formatting is done in the separate, platform agnostic methods.
The only exception is line break, because some of the ports use outputting methods which write line break on their own. This might be investigated and improved later.

This complements and corrects WriteStdOut() behavior after e8345e1ed5d0f02522606af419133e24c6260c78