Editor: Allow room objects to be locked like GUI controls
AGS allows the user to lock GUI controls, but not room objects, whose
placements can be just as intricate. This implements locking for room
objects with consistent appearance and behaviour to locked GUI controls.
Locked objects can't be moved by mouse or keyboard and have an X in the
centre. The unused flag, OBJF_DELETED has been repurposed as
OBJF_LOCKED.