Compiler: Bind negation operator closely with literals
Operands on the right of an expression are now checked to see whether
someone is performing an operation on a negated value (e.g. `int x = 4 +
- 3`). Operands on the left are now checked to see whether they are a
negated literal. If so, this information is passed through to the point
of variable access so that the literal is parsed with its sign intact.
Additionally, the lone call to `accept_literal_value` was removed,
unifying the parsing of literals and constants.