Engine: remap unsafe write filepaths for new games too
This partially reverts the originally suggested behavior when the unsafe filepath for writing operations caused new games fail to open file. User feedback indicated that this change was causing confusion, and also required fixing scripts in the WIP projects after updating to the new version.

Since filepath remapping is going to be featured in the minor version update, it is expected that users should be able to update their projects without such hassle.