Engine: limit written config options to those set by setup program
Because user config may now be written into different file (e.g. in user documents path), and the config file in game's folder is considered "default" one, it would not be wise to copy all of the contents to user file, since default one may contain options that setup program does not (or should not) modify, or obsolete options from old versions of the engine.