Engine: GfxDriver's bitmap convertion should not delete original bitmap
ConvertBitmapToSupportedColourDepth() method of the IGraphicsDriver interface was used in such way that it is supposed to implicitly delete source bitmap if conversion was made. Not only such requirement for interface implementations stands against common sense, but may be very inconvenient when there is a need to keep original bitmap.

This commit removes bitmap deletion from the implementations of IGraphicsDriver::ConvertBitmapToSupportedColourDepth() method. To keep code changes to minimum, a helper function is used as replacement, which explicitly deletes source bitmap if graphics driver's method returned a new one.