Engine: revamped dynamic custom properties implementation
This makes a clear distinction between design-time custom property values and values set at runtime. They are stored in separate maps, and runtime values are only added if related SetProperty() function was called in script. When GetProperty() is called runtime map is searched first, if property is not found design-time map is searched. Only runtime values are written to game saves.