Engine: moved screen height and colour depth to savegame description
This does not change save data format itself, rather will interpret first two variables from the (previously) "game state" part as two additional variables at the end of the "save description" part. These variables are not part of game state, they are used to describe graphics mode in which the save was made, for compatibility test.