Common, Engine: moved the rest of game loading to Common library
The new main reading function is ReadGameData() now in the main_game_file unit. It requires LoadedGameEntities struct as parameter, to where it loads all the data. Therefore, game data reading is now independent from the engine and does not refer to global objects anymore.
Compatibility conversions are performed by UpdateGameData() which is supposed to be called after ReadGameData() returned with success.
InitGameState() takes LoadedGameEntities as parameter too, which it uses to finish game initialization.