Support custom capabilities list in the game data
This adds an optional list of arbitrary string tags into the game data header. These could be used by alternative versions of AGS to indicate their own game formats, because the basic data format number is not suitable for that purpose. They may also be taken into use for some needs of the "main" version in the future.