Engine: implemented optional vertical offset for built-in font types
The root of the issue is in the old (and possibly altered) true fonts library AGS is using. This makes AGS do not respect descend (lower part) of particular TTF fonts, making them draw lower than required. The solution is to add property that would alter font rendering's vertical offset.

On another hand, this property may come useful even after we switch to the contemporary font library, for the opposite backwards-compatible effect.

While the property is mainly suggested for TTF fonts, it affects both TTF and WFN.