Engine: provided backup variant for the InventoryScreen() resources
InventoryScreen() is a deprecated script function that displays hard-coded inventory window. It was used in the very old versions of AGS as a default way to work with inventory, but is no longer recommended. But we need to make sure it at least runs if called from script.
Implementation demands certain resources with hard-coded IDs to be present in the game data (text message, and three sprites). This fix provides backup variant if the game did not have these.