Editor: support conversion in Japanese and similar ANSI codepages
This fixes situation when the result of Unicode-to-ANSI conversion produces more that one bytes per character (as for Japanese and similar locales).

I had to rewrite some parts of the DataWriter, because when dealing with data size limits, it imposed them as character limit on Unicode strings, rather than actually byte limit on final conversion result.

For same reasons I had to rewrite one string encryption routine, because it was encrypting Unicode string rather than final ANSI byte array.