Common, Engine: use new linespacing rule when calculating text position
With the new logic font height is only directly used when you need to know the size of space single line of text takes.

When drawing text line by line, use linespacing parameter to advance Y coordinate.

When in the need of total height of multiple lines of text, use helper function getheightoflines(), which internally calculates the height as (linespacing * (numlines-1) + fontheight).