Error CS0219 in project AGS.CScript.Compiler, file CodeBlockCompiler.cs
Variable "setEqualTo" is assigned, but value is never used. I quite simply removed the variable assignment. We still need to call the function, as quite lot is happening in the code behind.

Variable "memberName" is assigned, but value is never used. There are several TODOs in the code surrounding this variable. I simply commented the variable in case of future needs.