Environment.OSVersion output correct OS
Environemnet.OSVersion wouldn't output correct OS version on Windows 8.1 and Windows 10. This is solved by a adding a manifest file with metadata that tells the application at build time that the OS in question are supported.

Microsoft do not actually recommend using Environemnet.OSVersion, but it works just fine  and they haven't provided alternatives that doesn't involve invoking native code or doing registry lookups, which I'm trying to avoid.