Improved .NET run time check
I removed the old run time check and replaced it with something simpler. It is still not perfect however..

So the most conventional way of obtaining the .NET run time is using "Environment.Version", however because of an odd design choice by Microsoft this is not a good solution. The problem is that all .NET run times from v4.0 and newer is bundled together in the same directory and even  assembly. The result of this is that "Environment.Version" will always return "v4.0" on run time V4.0 and greater. The recommended solution provided by Microsoft is doing registry queries, however this is obviously not very cross platform, code that does registry look ups also have a tendency to become ugly so we want to avoid this at all costs.

Since the .NET API doesn't provide us a good way to solve this issue we have to resort to a minor hack. What I've done is exectued a line of code that grabs the .dll that is responsible for hosting the run-time "mscorlib.dll", and check which version of the run time it has loaded internally. I've tested this on multiple .NET versions for Windows 10 and Ubuntu and it works great. Credit for this clever solution goes to [James L. on stackoverflow](http://stackoverflow.com/questions/8517159/how-do-i-detect-at-runtime-that-net-version-4-5-is-currently-running-your-code)

I also removed all traces of Windows specific code in StatisticsSenderComponent.cs