Engine: fixed Game.TranslationFilename return value
According to the manual (and older engine code) transFileName string should be empty if default language is used. That is unknown if there ever was actual support for "default.tra" file, because the original code never called init_translation() with an empty language name, it was always checked before the call.