Engine: implemented new save/load data routines
This implements the data save and restoration routines for new save format. They are based on old ones, but have better data grouping, no redundant data saved (such as static data that never changes during the course of the game) and additional assertions to help with debugging in case of future format updates.