Scripting: Fixed get_friendly_name past EOF
Modified the behaviour of get_friendly_name so that it more closely
matches the behaviour it replaced (get_name).

There are some areas in the compiler where it assumes the next symbol
is valid to print. These are errors in the compiler's logic, but whereas
get_name would give reasonable output, "const (null)", get_friendly_name
causes a crash.

Note: This was likely because sprintf dealt with a null c-style string
pointer and output "(null)".

This new behaviour isn't perfect, but it's better than hard crashes when
there are errors in the compiler logic. When there is a problem getting
the name of a symbol, it will now be shown as "(invalid symbol)", and we
can fix these issues as we find them.