Editor: hotfix script editor cannot undo after save
I honestly do not understand why, but Editor keeps Scintilla component in Read-only state most of the time, manually toggling it when the user types/pastes text.
Scintilla checks for ReadOnly when asked if it can Undo/Redo and may reply negative even when there are entries in the Undo buffer.

This fix is a hack, that similarily turns ReadOnly mode off before asking Scintilla if Undo is allowed.