Editor: reimplemented room item design-time settings as a class
Previously Visible and Locked settings of a room item were implemented as two separate lists of names, and presence of a name in a list would mean that corresponding setting is enabled for an item.
This commit reimplements these into a dictionary of configs named DesignTimeProperties that should be more convenient to work with and easier to ammend.

NOTE: because of the way item update is done, this code change will crash the editor whenever an item is added, removed or renamed, but this will be fixed by the following changes.