Engine: fixed dynamic array releasing managed items on forced Dispose
Explanation follows:

As of now, managed objects in AGS has two Dispose modes: forced and non-forced. The meaning of "forced" dispose is to simply detach the object from the managed system, breaking any links and references to other managed objects or game resources, instead of disposing these all along.
For example, DynamicSprite does not remove and delete an actual bitmap from the spriteset when ordered to be forcefully disposed.
The main case when the "forced" Dispose is used is savedgame restoration process.

The problem was that DynamicArray class was ignoring the "forced" flag, and continued to release its managed items in a default way, causing them to dispose their resources in turn. This could lead to some resources deallocated when they should not have to be.

This bug seems to be very old and was not pinpointed until recently because in practice it leads to errors only if DynamicArray is assigned an earlier index in the managed pool than all or some of its elements and its Dispose is called before theirs, and that made it not common in games.