Switching between windowed & fullscreen now works.
Previously we just called Allegro's set_gfx_mode() for each mode change, and
recreated the OpenGL context, but this led to a corrupted image. As of this
commit the switch to fullscreen is done by setting the _NET_WM_STATE_FULLSCREEN
property on the "managed" X11 Window created by Allegro. The OpenGL context
stays intact this way. This technique should work with modern window managers,
although has only minimal testing at this point in time.

One potential pitfall of this approach is that since we only call set_gfx_mode()
once, that is the only time the desktop video mode will be set. I'm not yet sure
if that's an actual problem, or just a benign difference with respect to the
software renderer.

A benefit of this approach is that keys such as alt-tab and volume control
multimedia keys now DTRT when in fullscreen mode.