Engine: fixed software render's dirty rectangles function for 16-bit games
I did not realize Allegro does not support stretch_blit between bitmaps with different colour depths, so when drawin 16-bit games on 32-bit virtual screen the room backgrounds were broken.
Furthermore, we should not be using any kind of transform (other than linear offsets) in the dirty rects function anyway because for software renderer it is imperative that the room background is prepared untransformed for drawing objects and characters on it.