Common, Editor: implemented "real" sprite resolution type
This deprecates sprite's "resolution" setting and defaults it to "Real resolution", which means that sprite is displayed in same size in all circumstances.
Upon loading or importing sprite data, sprites are "promoted" to real resolution when possible, that is if it matches game resolution type.

I decided to mark the sprites of variable resolution with special flag upon updating old data instead of enforcing new data have some flag set, because this setting should really have been optional from the beginning.