Engine: removed skipping render if screen is faded out
There is a number of things wrong with this.
1) This will skip any engine's utility overlay, such as debug info or fps counter, which should not happen since "fade out" is an in-game effect.
2) For 3D renderers (e.g. Direct3D, OpenGL) this will skip vsync, causing game to run faster when fade out.
3) Existing code also causes plugins to miss number of events (coded as pseudo-sprites) if game is run with 3D renderer.

I would also reimplement this skip during scene setup (construct_virtual_screen) where skip adding in-game sprites to the render lists, but apparently drawing hooks are the only way for plugins to process on every game tick at the moment, so more accurate approach is required.