add freetype 2.1.3 to libsrc
freetype 2.1.3 is the version shipped with the windows version of AGS, and the
one the custom patched alfont library targets.
unlike more recent freetype versions, it works as expected with games of the
"maniac mansion mania" series of mini-adventures, and others that use "LucasFan"
and similar fonts.

archive found on
http://ftp.math.utah.edu/pub/freetype/freetype-2.1.3.tar.gz

$ sha256sum freetype-2.1.3.tar.gz
b9b2e43195bc79509dabdcbe823e09cbbac201b55981617e2b2804ecf094e21e

the freetype sources used with the windows port have some modifications to the
Makefiles, which will be added in the next commit.