audio: simplify clip->destroy behaviour
Previously 'poll would destroy if 'done' && 'multithreaded'. 'play' would destroy/delete on error. This lead to some issues where clips were assigned to channel and played, but then not checked if they succeeded, leading to pointers to freed memory.

Destroy/delete of clips is now up to the main thread. Soundclips will not attempt to delete themselves or destroy during poll. Code calling 'play' will clean up clip if it is not assigned to a channel (where it would get destroyed upon replacement).

Add a state to soundClip to encapsulate initialising, playing, paused and done. With these extra checks, we don't always need to check if clip is playing before accessing it.