Engine: removed incorrect ScriptCamera and ScriptViewport save format
This serialization format was added in expectation that ScriptCamera and ScriptViewport types are going to be self-managing, but now this appears to be a mistake, because internal game state must keep an ownership of these at all times and Script objects are going to be mere references (holding camera & viewports ID).
This is important to note that this old format was not working correctly anyway, because it was saving and restoring same data for each shared reference, while all of those were pointing same internal object.

I changed script object's type string to ensure that there's a clear error message if engine tries to read a savegame made with the previous format.
I believe it is not necessary to support previous format in this case, because it's unlikely that there were already real games published using new script commands.